Backend/requirements.txt
Flask>=2.1.0
Flask-Cors>=3.0.10
PyJWT>=2.6.0
pyserial>=3.5
werkzeug>=2.1.0
